style(reporter): rename files with TestFilePathResolver
