+++ packages/reporter/src/utils/index.ts
@@ -3,5 +3,5 @@ export { IssuesGrouper } from './issues-grouper';
 export {
   TEST_FILE_PATH_RESOLVER,
   TestFilePathResolver
-} from './test-file-path-resolver';
-export { JestTestFilePathResolver } from './jest-test-file-path-resolver';
+} from './TestFilePathResolver';
+export { JestTestFilePathResolver } from './JestTestFilePathResolver';
