+++ packages/reporter/src/utils/register.ts
@@ -1,5 +1,5 @@
-import { TEST_FILE_PATH_RESOLVER } from './test-file-path-resolver';
-import { JestTestFilePathResolver } from './jest-test-file-path-resolver';
+import { TEST_FILE_PATH_RESOLVER } from './TestFilePathResolver';
+import { JestTestFilePathResolver } from './JestTestFilePathResolver';
 import { container } from 'tsyringe';
 
 container.registerSingleton(TEST_FILE_PATH_RESOLVER, JestTestFilePathResolver);
